перенес функцию
