+++ src/structs/structs.h
@@ -436,6 +436,12 @@ class ItemName {
 };
 
 }
+//спецпроцедуры
+int exchange(CharData *ch, void *me, int cmd, char *argument);
+int horse_keeper(CharData *ch, void *me, int cmd, char *argument);
+int torc(CharData *ch, void *me, int cmd, char *argument);
+int mercenary(CharData *ch, void * /*me*/, int cmd, char *argument);
+int shop_ext(CharData *ch, void *me, int cmd, char *argument);
 
 #endif // STRUCTS_STRUCTS_H_
